refactor: Move  a bunch of logic down a level to avoid weird component boundaries
