+++ lib/DocumentFilter/DocumentFilter.js
@@ -31,11 +31,6 @@ const DocumentFilter = ({
     categoryValuesToUse = atTypeValues;
   }
 
-  // categoryValues are only passed for SupplementaryDocumentFilter
-  const [editingFilters, setEditingFilters] = useState(false);
-  const openEditModal = () => setEditingFilters(true);
-  const closeEditModal = () => setEditingFilters(false);
-
   // Due to how filters are handled within SearchAndSortQuery the filter string needs to be parsed back into a usual object
   const parseQueryString = (filterArray) => {
     if (filterArray?.length) {
@@ -72,42 +67,6 @@ const DocumentFilter = ({
 
   const parsedFilterData = parseQueryString(activeFilters?.[filterName] || []);
 
-  const handleSubmit = (values) => {
-    // In order to convert the form values into the shape for them to be deparsed we do the inverse of the above
-    // Adding a || operator between all elements of the filters array and a && operator between all elements of the nested arrays
-    // With special logic to ensure that operators are not added infront of the first elements of any arrays, to ensure no grouping errors
-    const kiwtQueryShape = values?.filters?.reduce((acc, curr, index) => {
-      let newAcc = [...acc];
-
-      if (index !== 0) {
-        newAcc = [...newAcc, '||'];
-      }
-
-      newAcc = [
-        ...newAcc,
-        curr.rules.reduce((a, c, i) => {
-          return [
-            ...a,
-            i !== 0 ? '&&' : null, // Don't group on first entry
-            c,
-          ].filter(Boolean);
-        }, []),
-      ];
-
-      return newAcc;
-    }, []);
-
-    filterHandlers.state({
-      ...activeFilters,
-      [filterName]: [
-        // Currently the deparse function returns a query string containing whitespace which leads to grouping errors
-        // This regex removes all whitespace from the querystring
-        deparseKiwtQueryFilters(kiwtQueryShape),
-      ],
-    });
-    closeEditModal();
-  };
-
   return (
     <Accordion
       closedByDefault
@@ -129,16 +88,12 @@ const DocumentFilter = ({
         </Layout>
       )}
       <DocumentFilterForm
+        activeFilters={activeFilters}
         categoryValues={categoryValuesToUse}
-        editingFilters={editingFilters}
+        filterHandlers={filterHandlers}
         filterModalProps={filterModalProps}
         filterName={filterName}
         filters={parsedFilterData}
-        handlers={{
-          closeEditModal,
-          openEditModal,
-        }}
-        onSubmit={handleSubmit}
       />
     </Accordion>
   );
