+++ lib/DocumentFilter/DocumentFilter.test.js
@@ -14,30 +14,10 @@ import DocumentFilter from './DocumentFilter';
 const testPFD = jest.fn();
 
 jest.mock('./DocumentFilterForm', () => ({
-  editingFilters,
   filters,
-  handlers,
-  onSubmit
 }) => {
   return (
     <div>
-      <StripesButton onClick={() => handlers.openEditModal()}>OPEN</StripesButton>
-      <StripesButton onClick={() => handlers.closeEditModal()}>CLOSE</StripesButton>
-      {editingFilters ? 'editingFilters' : ''}
-      <StripesButton
-        onClick={() => onSubmit({
-          filters: [{
-            rules: [{
-              comparator: '==',
-              path: 'docs.name',
-              value: 'do the test'
-            }]
-          }]
-        })
-        }
-      >
-        SUBMIT
-      </StripesButton>
       <StripesButton onClick={() => testPFD(filters)}>Test parsed filter data</StripesButton>
       DocumentFilterForm
     </div>
@@ -103,7 +83,7 @@ describe('DocumentFilter', () => {
     // Describe clicking close
       // Test editingFilters is false
 
-  // Open editingFilters
+  // Open editingFilters -- MOVE THESE TESTS ONE LEVEL DOWN
   // Click submit
     // test stateMock.calls[0][0] is {docs: ["docs.name==do the test"]}
     // test editingFIlters is false
