+++ lib/DocumentFilter/DocumentFilterForm.js
@@ -1,19 +1,60 @@
 import PropTypes from 'prop-types';
 import { Button } from '@folio/stripes/components';
 import arrayMutators from 'final-form-arrays';
-import { FormModal } from '@k-int/stripes-kint-components';
+import { deparseKiwtQueryFilters, FormModal } from '@k-int/stripes-kint-components';
 import { FormattedMessage } from 'react-intl';
 import DocumentFilterFieldArray from './DocumentFilterFieldArray';
+import { useState } from 'react';
 
 const DocumentFilterForm = ({
+  activeFilters,
   categoryValues,
-  editingFilters,
+  filterHandlers,
   filterModalProps,
   filterName,
   filters,
-  handlers: { openEditModal, closeEditModal },
-  onSubmit,
 }) => {
+  // categoryValues are only passed for SupplementaryDocumentFilter
+  const [editingFilters, setEditingFilters] = useState(false);
+  const openEditModal = () => setEditingFilters(true);
+  const closeEditModal = () => setEditingFilters(false);
+
+  const handleSubmit = (values) => {
+    // In order to convert the form values into the shape for them to be deparsed we do the inverse of the above
+    // Adding a || operator between all elements of the filters array and a && operator between all elements of the nested arrays
+    // With special logic to ensure that operators are not added infront of the first elements of any arrays, to ensure no grouping errors
+    const kiwtQueryShape = values?.filters?.reduce((acc, curr, index) => {
+      let newAcc = [...acc];
+
+      if (index !== 0) {
+        newAcc = [...newAcc, '||'];
+      }
+
+      newAcc = [
+        ...newAcc,
+        curr.rules.reduce((a, c, i) => {
+          return [
+            ...a,
+            i !== 0 ? '&&' : null, // Don't group on first entry
+            c,
+          ].filter(Boolean);
+        }, []),
+      ];
+
+      return newAcc;
+    }, []);
+
+    filterHandlers.state({
+      ...activeFilters,
+      [filterName]: [
+        // Currently the deparse function returns a query string containing whitespace which leads to grouping errors
+        // This regex removes all whitespace from the querystring
+        deparseKiwtQueryFilters(kiwtQueryShape),
+      ],
+    });
+    closeEditModal();
+  };
+
   return (
     <>
       <Button onClick={openEditModal}>
@@ -33,7 +74,7 @@ const DocumentFilterForm = ({
           ...filterModalProps
         }}
         mutators={{ ...arrayMutators }}
-        onSubmit={onSubmit}
+        onSubmit={handleSubmit}
       >
         <DocumentFilterFieldArray
           categoryValues={categoryValues}
@@ -46,7 +87,6 @@ const DocumentFilterForm = ({
 
 DocumentFilterForm.propTypes = {
   categoryValues: PropTypes.arrayOf(PropTypes.object),
-  editingFilters: PropTypes.bool,
   filterModalProps: PropTypes.object.isRequired,
   filterName: PropTypes.string.isRequired,
   filters: PropTypes.arrayOf(PropTypes.object),
@@ -54,6 +94,5 @@ DocumentFilterForm.propTypes = {
     closeEditModal: PropTypes.func.isRequired,
     openEditModal: PropTypes.func.isRequired,
   }),
-  onSubmit: PropTypes.func,
 };
 export default DocumentFilterForm;
